[refactor]: Header 컴포넌트 구조 개선
- BaseHeader 컴포넌트로 공통 레이아웃 로직 통합
- 헤더 관련 상수 분리
- 스크롤 관련 로직 usePostScroll 커스텀 훅으로 분리
