+++ src/components/layout/Header/HeaderLogo.tsx
@@ -1,0 +1,9 @@
+import Link from "next/link";
+
+export default function HeaderLogo() {
+  return (
+    <Link href="/" className="shrink-0">
+      <h3 className="text-3xl font-extrabold">chann-log</h3>
+    </Link>
+  );
+}
